test(agent): split selected SWM lifecycle specs
