+++ packages/agent/test/selected-swm-continuation.test.ts
@@ -36,811 +36,25 @@ import {
 } from '../src/sync/requester/page-fetch.js';
 import { estimateQuadHeapBytes } from '../src/sync/memory-telemetry.js';
 import { DURABLE_DATA_SYNC_SESSION_TTL_MS } from '../src/sync/durable-session.js';
-
-const PEER = '12D3KooWSelectedCompleteSwmProvider';
-
-const DKG = 'http://dkg.io/ontology/';
-const XSD_INTEGER = 'http://www.w3.org/2001/XMLSchema#integer';
-
-function snapshotManifest(contextGraphId: string, count: number): {
-  meta: Quad[];
-  payloadByRef: Map<string, Quad[]>;
-} {
-  const metaGraph = contextGraphWorkspaceMetaGraphUri(contextGraphId);
-  const meta: Quad[] = [];
-  const payloadByRef = new Map<string, Quad[]>();
-  for (let index = 0; index < count; index += 1) {
-    const payload: Quad[] = [{
-      subject: `urn:selected-swm:${index}`,
-      predicate: 'http://schema.org/value',
-      object: `"${index}"`,
-      graph: '',
-    }];
-    const digest = workspacePublicQuadsDigest(payload);
-    const subject = `urn:selected-swm-manifest:${index}`;
-    meta.push(
-      {
-        subject,
-        predicate: `${DKG}publicQuadsDigest`,
-        object: `"${digest}"`,
-        graph: metaGraph,
-      },
-      {
-        subject,
-        predicate: `${DKG}publicQuadsCount`,
-        object: '"1"',
-        graph: metaGraph,
-      },
-    );
-    payloadByRef.set(digest, payload);
-  }
-  return { meta, payloadByRef };
-}
-
-function graphBackedManifest(contextGraphId: string): ReturnType<typeof snapshotManifest> {
-  const metaGraph = contextGraphWorkspaceMetaGraphUri(contextGraphId);
-  const kaUal = 'did:dkg:testnet:20430/0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/1';
-  const assertionVersion = 1;
-  const shareOperationId = 'graph-backed-selected-op';
-  const operationSubject = `urn:dkg:share:${contextGraphId}:${shareOperationId}`;
-  const headSubject = `${kaUal}#dkg-swm-head`;
-  const assertionGraph = knowledgeAssetLayerGraphUri(
-    contextGraphId,
-    MemoryLayer.SharedWorkingMemory,
-    createGraphKnowledgeAssetScope(kaUal, assertionVersion),
-  );
-  const snapshotGraph = `did:dkg:context-graph:${encodeURIComponent(contextGraphId)}`
-    + `/_shared_memory_snapshots/_/${encodeURIComponent(shareOperationId)}/ka`;
-  const payload: Quad[] = [{
-    subject: 'urn:selected-graph-backed',
-    predicate: 'http://schema.org/value',
-    object: '"missing-from-response"',
-    graph: '',
-  }];
-  const digest = workspacePublicQuadsDigest(payload);
-  const meta: Quad[] = [
-    ...generateKnowledgeAssetShareMetadata({
-      shareOperationId,
-      contextGraphId,
-      kaUal,
-      assertionVersion,
-      publicTripleCount: payload.length,
-      privateTripleCount: 0,
-      publisherPeerId: PEER,
-      timestamp: new Date(0),
-    }, metaGraph),
-    {
-      subject: operationSubject,
-      predicate: `${DKG}publicQuadsDigest`,
-      object: `"${digest}"`,
-      graph: metaGraph,
-    },
-    {
-      subject: operationSubject,
-      predicate: `${DKG}publicSnapshotGraph`,
-      object: snapshotGraph,
-      graph: metaGraph,
-    },
-    {
-      subject: headSubject,
-      predicate: `${DKG}contentScopeVersion`,
-      object: `"${GRAPH_KA_CONTENT_SCOPE_VERSION}"^^<${XSD_INTEGER}>`,
-      graph: metaGraph,
-    },
-    { subject: headSubject, predicate: `${DKG}kaUal`, object: kaUal, graph: metaGraph },
-    {
-      subject: headSubject,
-      predicate: `${DKG}assertionVersion`,
-      object: `"${assertionVersion}"^^<${XSD_INTEGER}>`,
-      graph: metaGraph,
-    },
-    { subject: headSubject, predicate: `${DKG}assertionGraph`, object: assertionGraph, graph: metaGraph },
-    {
-      subject: headSubject,
-      predicate: `${DKG}shareOperationId`,
-      object: `"${shareOperationId}"`,
-      graph: metaGraph,
-    },
-  ];
-  return { meta, payloadByRef: new Map() };
-}
-
-function cleanDurableResult(): SharedMemorySyncResult {
-  return {
-    insertedTriples: 0,
-    fetchedMetaTriples: 0,
-    fetchedDataTriples: 0,
-    insertedMetaTriples: 0,
-    insertedDataTriples: 0,
-    bytesReceived: 0,
-    resumedPhases: 0,
-    timedOutPhases: 0,
-    completedPhases: 0,
-    checkpointAdvances: 0,
-    emptyResponses: 0,
-    droppedDataTriples: 0,
-    failedPeers: 0,
-    failedPhases: 0,
-    deniedPhases: 0,
-    backoffWorthyFailures: 0,
-    deferredBackpressure: 0,
-    snapshotPlaneIncomplete: 0,
-    replayPhaseBytesReceived: 0,
-    snapshotPhaseBytesReceived: 0,
-  };
-}
-
-function result(
-  contextGraphId: string,
-  snapshotsResolved: number,
-  snapshotsTotal: number,
-  options: {
-    completed?: boolean;
-    deferredBackpressure?: number;
-    insertedDataTriples?: number;
-  } = {},
-): SharedMemorySyncResult {
-  const completed = options.completed ?? snapshotsResolved === snapshotsTotal;
-  const swmCoverage: SwmSnapshotCoverage = {
-    contextGraphId,
-    peerIdSuffix: PEER.slice(-8),
-    snapshotsResolved,
-    snapshotsTotal,
-    manifestComplete: true,
-    descriptorsAuthoritative: true,
-    missingCount: snapshotsTotal - snapshotsResolved,
-    missingSample: [],
-    materializationFailures: 0,
-  };
-  return {
-    insertedTriples: options.insertedDataTriples ?? snapshotsResolved,
-    fetchedMetaTriples: 0,
-    fetchedDataTriples: 0,
-    insertedMetaTriples: 0,
-    insertedDataTriples: options.insertedDataTriples ?? snapshotsResolved,
-    bytesReceived: 0,
-    resumedPhases: 0,
-    // Production voluntary snapshot yield: the responder is healthy, our
-    // local round budget ended with refs outstanding.
-    timedOutPhases: 0,
-    completedPhases: completed ? 1 : 0,
-    checkpointAdvances: snapshotsResolved > 0 ? 1 : 0,
-    emptyResponses: 0,
-    droppedDataTriples: 0,
-    failedPeers: 0,
-    failedPhases: completed ? 0 : 1,
-    deniedPhases: 0,
-    backoffWorthyFailures: 0,
-    deferredBackpressure: options.deferredBackpressure ?? 0,
-    snapshotPlaneIncomplete: completed ? 0 : 1,
-    replayPhaseBytesReceived: 0,
-    snapshotPhaseBytesReceived: 0,
-    swmCoverage,
-  };
-}
-
-function merge(
-  summary: SharedMemorySyncResult,
-  part: SharedMemorySyncResult,
-): SharedMemorySyncResult {
-  return {
-    ...part,
-    insertedTriples: summary.insertedTriples + part.insertedTriples,
-    insertedDataTriples: summary.insertedDataTriples + part.insertedDataTriples,
-    deferredBackpressure:
-      (summary.deferredBackpressure ?? 0) + (part.deferredBackpressure ?? 0),
-  };
-}
-
-function selectedUnit(
-  contextGraphId: string,
-  initialResult: SharedMemorySyncResult,
-  run: () => Promise<SharedMemorySyncResult>,
-  metadata: {
-    readonly initial?: SelectedSwmMetaContinuation;
-    readonly afterRun?: () => SelectedSwmMetaContinuation;
-  } = {},
-) {
-  const completeMetadata: SelectedSwmMetaContinuation = {
-    progress: undefined,
-    generation: 0,
-    completed: true,
-  };
-  return {
-    work: {
-      contextGraphId,
-      lane: 'shared_memory' as const,
-      operationId: contextGraphId,
-      run: async () => ({
-        result: await run(),
-        metadataContinuation: metadata.afterRun?.() ?? completeMetadata,
-      }),
-    },
-    initialRound: {
-      result: initialResult,
-      metadataContinuation: metadata.initial ?? completeMetadata,
-    },
-  };
-}
-
-interface SelectedProviderSelectionAgent {
-  started: boolean;
-  config: {
-    syncOnConnect: boolean;
-    syncSharedMemoryOnConnect: boolean;
-    syncContextGraphs: string[];
-    rfc64PublicCatalogBootstrap: {
-      acceptedPublicPolicies: Array<{ completeSwmProviders: string[] }>;
-    };
-  };
-  networkAdmissionCoordinator: { isAcceptedPeer: (peerId: string) => boolean };
-  syncingPeers: Set<string>;
-  knownCorePeerIds: Set<string>;
-  knownCorePeerIdsV2: Set<string>;
-  skippedNoSyncPeers: Set<string>;
-  lastSuccessfulSyncAt: Map<string, number>;
-  lastSyncProgressAt: Map<string, number>;
-  syncReconcilerBackoff: Map<string, unknown>;
-  selectedSwmRetryRequiredPeers: Set<string>;
-  getPeerProtocols: () => Promise<string[]>;
-  planSharedMemorySyncContextGraphs: () => Promise<{
-    publicContextGraphIds: string[];
-    privateRecoverFromCurator: string[];
-    eligibleContextGraphIds: string[];
-  }>;
-  resolveRfc64CompleteSwmProviderPeerIdsV1: (contextGraphId: string) => string[];
-  syncFromPeerDetailed: () => Promise<number>;
-  refreshMetaSyncedFlags: () => Promise<void>;
-  discoverContextGraphsFromStore: () => Promise<number>;
-  syncSharedMemoryFromPeerDetailed: (
-    peerId: string,
-    contextGraphIds: readonly string[],
-  ) => Promise<SharedMemorySyncResult>;
-  syncSelectedSharedMemoryFromPeerDetailed: (
-    peerId: string,
-    contextGraphIds: readonly string[],
-    options: { selectedSwmPriority: true },
-  ) => Promise<{
-    kind: 'selected-shared-memory';
-    shared: SharedMemorySyncResult;
-    selectedScopeComplete: boolean;
-  }>;
-  log: { info: () => void; warn: () => void; debug: () => void };
-  getSelectedSwmMetaTransfers: () => SelectedSwmMetaTransferCoordinator;
-  closeSelectedSwmMetaTransfers: () => Promise<void>;
-}
-
-const callTrySyncFromPeer = LifecycleSyncMethods.prototype.trySyncFromPeer as unknown as (
-  this: SelectedProviderSelectionAgent,
-  remotePeer: string,
-  onSyncAccounting?: (outcome: { fresh: boolean; progress?: boolean }) => void,
-) => Promise<unknown>;
-
-interface AdmissionProbe {
-  readonly contextGraphId: string;
-  readonly selected: boolean;
-  readonly priority: number | undefined;
-  readonly event: 'start' | 'end';
-}
-
-interface SelectedSwmLifecycleHarnessOptions {
-  readonly contextGraphs: {
-    readonly public: string;
-    readonly private?: string;
-  };
-  readonly manifest: ReturnType<typeof snapshotManifest>;
-  readonly clock: {
-    readonly now: () => number;
-    readonly deadline: () => number;
-  };
-  readonly priorities?: Readonly<Record<string, number>>;
-  readonly onSnapshotRead?: (probe: {
-    readonly ref: string;
-    readonly publicAdmission: number;
-    readonly snapshotRead: number;
-  }) => void;
-  readonly beforeAdmissionRun?: (probe: {
-    readonly contextGraphId: string;
-    readonly publicAdmission: number;
-  }) => Promise<void> | void;
-  readonly onMetaFetch?: (probe: {
-    readonly fetch: number;
-    readonly requesterScope: string | undefined;
-    readonly maxAcceptedQuads: number | undefined;
-    readonly maxAcceptedHeapBytesEstimate: number | undefined;
-  }) => Promise<void> | void;
-  readonly metaContinuationLimits?: {
-    readonly rows: number;
-    readonly bytesEstimate: number;
-    readonly globalRows?: number;
-    readonly globalBytesEstimate?: number;
-  };
-  /** Deterministic metadata slices returned by consecutive selected passes. */
-  readonly metaPages?: readonly {
-    readonly quads: Quad[];
-    readonly resumedFromOffset: number;
-    readonly nextOffset: number;
-    readonly completed: boolean;
-    readonly timedOut: boolean;
-    readonly responderSessionStartedFresh?: boolean;
-  }[];
-  /** Number of aggregate-data calls that fail after metadata completed. */
-  readonly dataFailuresBeforeSuccess?: number;
-  /** Mirror the production verifier's empty-batch classification. */
-  readonly reportEmptyResponse?: boolean;
-  /** Override the aggregate data phase to exercise incomplete empty rounds. */
-  readonly dataPage?: {
-    readonly quads?: readonly Quad[];
-    readonly completed: boolean;
-    readonly timedOut: boolean;
-  };
-}
-
-interface SelectedSwmLifecycleAgentFixture {
-  config: {
-    syncContextGraphPriorities: Readonly<Record<string, number>>;
-    syncResponderSnapshotLimits?: {
-      global?: { rows?: number; bytesEstimate?: number };
-      local?: { rows?: number; bytesEstimate?: number };
-    };
-  };
-  selectedSwmRetryRequiredPeers: Set<string>;
-  store: OxigraphStore;
-  writeLocks: Map<string, Promise<void>>;
-  publicSnapshotStore: {
-    getSnapshot: (ref: string) => Promise<Quad[] | null>;
-    putSnapshot: (input: { digest: string }) => Promise<{ ref: string; byteLength: number }>;
-  };
-  listSubGraphs: () => Promise<string[]>;
-  createContextGraphSyncDeadline: () => number;
-  fetchSyncPages: (
-    ctx: unknown,
-    peerId: string,
-    contextGraphId: string,
-    includeSharedMemory: boolean,
-    phase: string,
-    graphUri: string,
-    deadline: number,
-    options?: SyncPageFetchOptions,
-  ) => Promise<{
-    quads: Quad[];
-    bytesReceived: number;
-    resumedFromOffset: number;
-    nextOffset: number;
-    checkpointKey: string;
-    completed: boolean;
-    timedOut: boolean;
-    responderSessionStartedFresh?: boolean;
-  }>;
-  getOrCreateSyncVerifyWorker: () => {
-    processSharedMemoryBatch: (dataQuads: Quad[], metaQuads: Quad[]) => Promise<{
-      verifiedData: Quad[];
-      verifiedMeta: Quad[];
-      totalFetchedDataQuads: number;
-      totalFetchedMetaQuads: number;
-      droppedDataTriples: number;
-      emptyResponses: number;
-      entityCreators: string[];
-    }>;
-  };
-  runContextGraphSyncWithBackpressure: (
-    ctx: unknown,
-    contextGraphId: string,
-    lane: string,
-    operationId: string,
-    work: () => Promise<SharedMemorySyncResult>,
-    admission: { priorityOverride?: number; selectedSwmPriority?: boolean },
-  ) => Promise<SharedMemorySyncResult>;
-  syncCheckpoints: Map<string, number>;
-  workspaceOwnedEntities: Map<string, Map<string, string>>;
-  invalidateListContextGraphsCache: () => void;
-  contextGraphMetaProjection: { markDirtyFromQuads: () => void };
-  oversizeTombstoneLog: { record: () => void };
-  log: { info: () => void; warn: () => void; debug: () => void };
-  syncSharedMemoryFromPeerDetailedExecution:
-    typeof LifecycleSyncMethods.prototype.syncSharedMemoryFromPeerDetailedExecution;
-}
-
-interface SelectedSwmLifecycleHarness {
-  readonly agent: SelectedSwmLifecycleAgentFixture;
-  readonly probes: {
-    readonly admissions: AdmissionProbe[];
-    readonly snapshotFetches: string[];
-    readonly publicAdmissions: () => number;
-    readonly snapshotReads: () => number;
-    readonly metaFetches: () => number;
-    readonly processedMetaBatches: readonly Quad[][];
-    readonly dataFetches: () => number;
-    readonly metaRequesterScopes: readonly (string | undefined)[];
-    readonly metaSinceBatchIds: readonly (string | undefined)[];
-    readonly metaReturnAcceptedPrefixOnRetryableTransportFailure: readonly boolean[];
-    readonly maxActiveAdmissions: () => number;
-  };
-  readonly close: () => Promise<void>;
-}
-
-type SyncSharedMemoryOptions = Parameters<
-  typeof LifecycleSyncMethods.prototype.syncSelectedSharedMemoryFromPeerDetailed
->[2];
-
-const callSelectedSharedMemoryFromPeerDetailed = (
-  agent: SelectedSwmLifecycleAgentFixture,
-  contextGraphIds: string[],
-  options: SyncSharedMemoryOptions,
-): Promise<{
-  kind: 'selected-shared-memory';
-  shared: SharedMemorySyncResult;
-  selectedScopeComplete: boolean;
-}> => {
-  const method = LifecycleSyncMethods.prototype.syncSelectedSharedMemoryFromPeerDetailed as unknown as (
-    this: SelectedSwmLifecycleAgentFixture,
-    remotePeerId: string,
-    ids: string[],
-    syncOptions: SyncSharedMemoryOptions,
-  ) => Promise<{
-    kind: 'selected-shared-memory';
-    shared: SharedMemorySyncResult;
-    selectedScopeComplete: boolean;
-  }>;
-  return method.call(agent, PEER, contextGraphIds, options);
-};
-
-const callSyncSharedMemoryFromPeerDetailed = async (
-  agent: SelectedSwmLifecycleAgentFixture,
-  contextGraphIds: string[],
-  options: SyncSharedMemoryOptions,
-): Promise<SharedMemorySyncResult> => (
-  (await callSelectedSharedMemoryFromPeerDetailed(agent, contextGraphIds, options)).shared
-);
-
-function createSelectedSwmLifecycleHarness(
-  options: SelectedSwmLifecycleHarnessOptions,
-): SelectedSwmLifecycleHarness {
-  const store = new OxigraphStore();
-  const admissions: AdmissionProbe[] = [];
-  const snapshotFetches: string[] = [];
-  let activeAdmissions = 0;
-  let maxActiveAdmissions = 0;
-  let publicAdmissions = 0;
-  let snapshotReads = 0;
-  let metaFetches = 0;
-  let dataFetches = 0;
-  const metaRequesterScopes: Array<string | undefined> = [];
-  const metaSinceBatchIds: Array<string | undefined> = [];
-  const metaReturnAcceptedPrefixOnRetryableTransportFailure: boolean[] = [];
-  const processedMetaBatches: Quad[][] = [];
-  const dateNow = vi.spyOn(Date, 'now').mockImplementation(options.clock.now);
-  let selectedSwmMetaTransfers: SelectedSwmMetaTransferCoordinator | undefined;
-
-  const agent: SelectedSwmLifecycleAgentFixture = {
-    config: {
-      syncContextGraphPriorities: options.priorities ?? {},
-      ...(options.metaContinuationLimits
-        ? {
-          syncResponderSnapshotLimits: {
-            local: {
-              rows: options.metaContinuationLimits.rows,
-              bytesEstimate: options.metaContinuationLimits.bytesEstimate,
-            },
-            ...(options.metaContinuationLimits.globalRows !== undefined
-              || options.metaContinuationLimits.globalBytesEstimate !== undefined
-              ? {
-                global: {
-                  rows: options.metaContinuationLimits.globalRows,
-                  bytesEstimate: options.metaContinuationLimits.globalBytesEstimate,
-                },
-              }
-              : {}),
-          },
-        }
-        : {}),
-    },
-    selectedSwmRetryRequiredPeers: new Set(),
-    store,
-    writeLocks: new Map(),
-    publicSnapshotStore: {
-      getSnapshot: async (ref) => {
-        snapshotReads += 1;
-        options.onSnapshotRead?.({ ref, publicAdmission: publicAdmissions, snapshotRead: snapshotReads });
-        return options.manifest.payloadByRef.get(ref) ?? null;
-      },
-      putSnapshot: async ({ digest }) => ({ ref: digest, byteLength: 0 }),
-    },
-    listSubGraphs: async () => [],
-    createContextGraphSyncDeadline: options.clock.deadline,
-    fetchSyncPages: async (
-      _ctx,
-      _peerId,
-      contextGraphId,
-      _includeSharedMemory,
-      phase,
-      _graphUri,
-      _deadline,
-      fetchOptions = {},
-    ) => {
-      const {
-        snapshotRef,
-        sinceBatchId,
-        requesterScope,
-        maxAcceptedQuads,
-        maxAcceptedHeapBytesEstimate,
-        returnAcceptedPrefixOnRetryableTransportFailure,
-      } = fetchOptions;
-      if (phase === 'snapshot') {
-        snapshotFetches.push(snapshotRef ?? 'missing-ref');
-        throw new Error('all snapshot fixtures should be cache hits');
-      }
-      if (phase === 'meta') {
-        metaFetches += 1;
-        metaRequesterScopes.push(requesterScope);
-        metaSinceBatchIds.push(sinceBatchId);
-        metaReturnAcceptedPrefixOnRetryableTransportFailure.push(
-          returnAcceptedPrefixOnRetryableTransportFailure === true,
-        );
-        await options.onMetaFetch?.({
-          fetch: metaFetches,
-          requesterScope,
-          maxAcceptedQuads,
-          maxAcceptedHeapBytesEstimate,
-        });
-        const planned = options.metaPages?.[metaFetches - 1];
-        if (planned) {
-          if (
-            maxAcceptedQuads !== undefined
-            && planned.quads.length > maxAcceptedQuads
-          ) {
-            const error = new SyncPageAccumulationLimitError(
-              'quads',
-              planned.quads.length,
-              maxAcceptedQuads,
-            );
-            error.responderSessionStartedFresh =
-              planned.responderSessionStartedFresh;
-            throw error;
-          }
-          const heapBytesEstimate = planned.quads.reduce(
-            (total, quad) => total + estimateQuadHeapBytes(quad),
-            0,
-          );
-          if (
-            maxAcceptedHeapBytesEstimate !== undefined
-            && heapBytesEstimate > maxAcceptedHeapBytesEstimate
-          ) {
-            const error = new SyncPageAccumulationLimitError(
-              'heap-bytes',
-              heapBytesEstimate,
-              maxAcceptedHeapBytesEstimate,
-            );
-            error.responderSessionStartedFresh =
-              planned.responderSessionStartedFresh;
-            throw error;
-          }
-          return {
-            ...planned,
-            bytesReceived: planned.quads.length,
-            checkpointKey: `${contextGraphId}:${phase}`,
-          };
-        }
-      }
-      if (phase === 'data') {
-        dataFetches += 1;
-        if (dataFetches <= (options.dataFailuresBeforeSuccess ?? 0)) {
-          throw new Error('simulated aggregate-data transport failure');
-        }
-        if (options.dataPage) {
-          return {
-            quads: [...(options.dataPage.quads ?? [])],
-            bytesReceived: options.dataPage.quads?.length ?? 0,
-            resumedFromOffset: 0,
-            nextOffset: options.dataPage.quads?.length ?? 0,
-            checkpointKey: `${contextGraphId}:${phase}`,
-            completed: options.dataPage.completed,
-            timedOut: options.dataPage.timedOut,
-          };
-        }
-      }
-      const quads = phase === 'meta' && contextGraphId === options.contextGraphs.public
-        ? options.manifest.meta
-        : [];
-      return {
-        quads,
-        bytesReceived: quads.length,
-        resumedFromOffset: 0,
-        nextOffset: quads.length,
-        checkpointKey: `${contextGraphId}:${phase}`,
-        completed: true,
-        timedOut: false,
-      };
-    },
-    getOrCreateSyncVerifyWorker: () => ({
-      processSharedMemoryBatch: async (dataQuads, metaQuads) => {
-        processedMetaBatches.push([...metaQuads]);
-        return {
-          verifiedData: dataQuads,
-          verifiedMeta: metaQuads,
-          totalFetchedDataQuads: dataQuads.length,
-          totalFetchedMetaQuads: metaQuads.length,
-          droppedDataTriples: 0,
-          emptyResponses: options.reportEmptyResponse
-            && dataQuads.length === 0
-            && metaQuads.length === 0
-              ? 1
-              : 0,
-          entityCreators: [],
-        };
-      },
-    }),
-    runContextGraphSyncWithBackpressure: async (
-      _ctx,
-      contextGraphId,
-      _lane,
-      _operationId,
-      work,
-      admission,
-    ) => {
-      activeAdmissions += 1;
-      maxActiveAdmissions = Math.max(maxActiveAdmissions, activeAdmissions);
-      if (contextGraphId === options.contextGraphs.public) publicAdmissions += 1;
-      const row = {
-        contextGraphId,
-        selected: admission.selectedSwmPriority === true,
-        priority: admission.priorityOverride,
-      };
-      admissions.push({ ...row, event: 'start' });
-      try {
-        await options.beforeAdmissionRun?.({ contextGraphId, publicAdmission: publicAdmissions });
-        return await work();
-      } finally {
-        admissions.push({ ...row, event: 'end' });
-        activeAdmissions -= 1;
-      }
-    },
-    syncCheckpoints: new Map(),
-    workspaceOwnedEntities: new Map(),
-    invalidateListContextGraphsCache: () => {},
-    contextGraphMetaProjection: { markDirtyFromQuads: () => {} },
-    oversizeTombstoneLog: { record: () => {} },
-    log: { info: () => {}, warn: () => {}, debug: () => {} },
-    syncSharedMemoryFromPeerDetailedExecution:
-      LifecycleSyncMethods.prototype.syncSharedMemoryFromPeerDetailedExecution,
-    getSelectedSwmMetaTransfers: () => {
-      selectedSwmMetaTransfers ??= new SelectedSwmMetaTransferCoordinator();
-      return selectedSwmMetaTransfers;
-    },
-    closeSelectedSwmMetaTransfers: async () => {
-      const transfers = selectedSwmMetaTransfers;
-      if (!transfers) return;
-      await transfers.close();
-      if (selectedSwmMetaTransfers === transfers) selectedSwmMetaTransfers = undefined;
-    },
-  };
-  return {
-    agent,
-    probes: {
-      admissions,
-      snapshotFetches,
-      publicAdmissions: () => publicAdmissions,
-      snapshotReads: () => snapshotReads,
-      metaFetches: () => metaFetches,
-      processedMetaBatches,
-      dataFetches: () => dataFetches,
-      metaRequesterScopes,
-      metaSinceBatchIds,
-      metaReturnAcceptedPrefixOnRetryableTransportFailure,
-      maxActiveAdmissions: () => maxActiveAdmissions,
-    },
-    close: async () => {
-      dateNow.mockRestore();
-      await agent.closeSelectedSwmMetaTransfers();
-      await store.close();
-    },
-  };
-}
-
-describe('shared-memory freshness classification', () => {
-  it('owns producer recovery, bounding, and final classification as one invariant', () => {
-    const contextGraphId = 'selected-freshness-boundary';
-    const incomplete = result(contextGraphId, 1, 2);
-    expect(classifySelectedSwmRoundFreshness(contextGraphId, incomplete)).toEqual({
-      recoverableSnapshotYieldFailures: 1,
-      recoverableMetadataContinuationYields: 0,
-      snapshotPlaneComplete: false,
-    });
-    expect(classifySelectedSwmRoundFreshness(contextGraphId, {
-      ...incomplete,
-      swmCoverage: {
-        ...incomplete.swmCoverage!,
-        contextGraphId: 'another-context-graph',
-      },
-    }).recoverableSnapshotYieldFailures).toBe(0);
-    expect(classifySelectedSwmRoundFreshness(contextGraphId, {
-      ...incomplete,
-      swmCoverage: {
-        ...incomplete.swmCoverage!,
-        descriptorsAuthoritative: false,
-      },
-    }).recoverableSnapshotYieldFailures).toBe(0);
-    expect(classifySelectedSwmRoundFreshness(contextGraphId, {
-      ...incomplete,
-      swmCoverage: {
-        ...incomplete.swmCoverage!,
-        materializationFailures: 1,
-      },
-    }).recoverableSnapshotYieldFailures).toBe(0);
-
-    const complete = result(contextGraphId, 2, 2);
-    expect(classifySelectedSwmRoundFreshness(contextGraphId, {
-      ...complete,
-      swmCoverage: {
-        ...complete.swmCoverage!,
-        descriptorsAuthoritative: false,
-      },
-    }).snapshotPlaneComplete).toBe(false);
-    const { descriptorsAuthoritative: _unknown, ...legacyCoverage } = complete.swmCoverage!;
-    expect(classifySelectedSwmRoundFreshness(contextGraphId, {
-      ...complete,
-      swmCoverage: legacyCoverage,
-    }).snapshotPlaneComplete).toBe(false);
-
-    const finalRaw = {
-      ...complete,
-      failedPhases: 1,
-      snapshotPlaneIncomplete: 1,
-    };
-    const resolved = applySelectedSwmFreshnessResolution(finalRaw, {
-      recoverableSnapshotYieldFailures: 1,
-    });
-    expect(classifySharedMemoryFreshness(resolved).phaseFailed).toBe(false);
-
-    expect(classifySharedMemoryFreshness({
-      ...finalRaw,
-      resolvedSnapshotPlaneIncomplete: 2,
-    }).phaseFailed).toBe(true);
-    expect(classifySharedMemoryFreshness({
-      ...finalRaw,
-      failedPhases: 2,
-      resolvedSnapshotPlaneIncomplete: 1,
-    }).phaseFailed).toBe(true);
-    expect(applySelectedSwmFreshnessResolution({
-      ...finalRaw,
-      failedPhases: Number.NaN,
-    }, {
-      recoverableSnapshotYieldFailures: 1,
-    }).resolvedSnapshotPlaneIncomplete).toBe(0);
-    expect(applySelectedSwmFreshnessResolution({
-      ...finalRaw,
-      snapshotPlaneIncomplete: -1,
-    }, {
-      recoverableSnapshotYieldFailures: 1,
-    }).resolvedSnapshotPlaneIncomplete).toBe(0);
-
-    const retainedMetadataYield = {
-      ...cleanDurableResult(),
-      timedOutPhases: 1,
-      metadataContinuationYields: 1,
-    };
-    expect(classifySelectedSwmRoundFreshness(
-      contextGraphId,
-      retainedMetadataYield,
-    ).recoverableMetadataContinuationYields).toBe(1);
-    const metadataResolved = applySelectedSwmFreshnessResolution(
-      {
-        ...retainedMetadataYield,
-        completedPhases: 2,
-      },
-      {
-        recoverableSnapshotYieldFailures: 0,
-        recoverableMetadataContinuationYields: 1,
-      },
-    );
-    expect(classifySharedMemoryFreshness(metadataResolved).timedOut).toBe(false);
-    expect(classifySharedMemoryFreshness({
-      ...retainedMetadataYield,
-      completedPhases: 2,
-      resolvedMetadataContinuationYields: 2,
-    }).timedOut).toBe(true);
-  });
-});
+import {
+  PEER,
+  callSelectedSharedMemoryFromPeerDetailed,
+  callSyncSharedMemoryFromPeerDetailed,
+  callTrySyncFromPeer,
+  cleanDurableResult,
+  createSelectedSwmLifecycleHarness,
+  graphBackedManifest,
+  merge,
+  result,
+  selectedUnit,
+  snapshotManifest,
+  type AdmissionProbe,
+  type SelectedProviderSelectionAgent,
+  type SelectedSwmLifecycleAgentFixture,
+  type SelectedSwmLifecycleHarness,
+  type SelectedSwmLifecycleHarnessOptions,
+  type SyncSharedMemoryOptions,
+} from './selected-swm-test-helpers.js';
 
 describe('selected RFC-64 SWM continuation', () => {
   it('re-enters admission promptly until incomplete coverage becomes complete', async () => {
@@ -1216,1617 +430,5 @@ describe('selected RFC-64 SWM continuation', () => {
     });
 
     expect(summary.continuationPasses).toBe(1);
-  });
-});
-
-describe('selected RFC-64 SWM lifecycle wiring', () => {
-  it('treats a clean selected graph with zero snapshot refs as explicit complete 0/0 coverage', async () => {
-    const publicCg = 'selected-empty-snapshot-plane';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest: snapshotManifest(publicCg, 0),
-      clock: { now: () => 1_000, deadline: () => 61_000 },
-      reportEmptyResponse: true,
-    });
-
-    try {
-      const selected = await callSelectedSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-      const summary = selected.shared;
-
-      expect(summary.swmCoverage).toEqual({
-        contextGraphId: publicCg,
-        peerIdSuffix: PEER.slice(-8),
-        snapshotsResolved: 0,
-        snapshotsTotal: 0,
-        manifestComplete: true,
-        descriptorsAuthoritative: true,
-        missingCount: 0,
-        missingSample: [],
-        materializationFailures: 0,
-      });
-      expect(selected.selectedScopeComplete).toBe(true);
-      expect(summary.continuationPasses).toBe(0);
-      expect(harness.probes.publicAdmissions()).toBe(1);
-      expect(harness.agent.selectedSwmRetryRequiredPeers.has(PEER)).toBe(false);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('keeps non-empty graph-backed zero-ref metadata incomplete until its content is proven', async () => {
-    const publicCg = 'selected-graph-backed-zero-ref';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest: graphBackedManifest(publicCg),
-      clock: { now: () => 1_000, deadline: () => 61_000 },
-    });
-
-    try {
-      const selected = await callSelectedSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(selected.shared.fetchedMetaTriples).toBeGreaterThan(0);
-      expect(selected.shared.insertedMetaTriples).toBe(0);
-      expect(selected.shared.failedPhases).toBeGreaterThan(0);
-      expect(selected.shared.swmCoverage).toBeUndefined();
-      expect(selected.selectedScopeComplete).toBe(false);
-      expect(harness.agent.selectedSwmRetryRequiredPeers.has(PEER)).toBe(true);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('keeps an empty selected graph incomplete when its data phase times out', async () => {
-    const publicCg = 'selected-empty-data-timeout';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest: snapshotManifest(publicCg, 0),
-      clock: { now: () => 1_000, deadline: () => 61_000 },
-      reportEmptyResponse: true,
-      dataPage: {
-        quads: [],
-        completed: false,
-        timedOut: true,
-      },
-    });
-
-    try {
-      const selected = await callSelectedSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(selected.shared.swmCoverage).toBeUndefined();
-      expect(selected.shared.timedOutPhases).toBeGreaterThan(0);
-      expect(selected.selectedScopeComplete).toBe(false);
-      expect(harness.agent.selectedSwmRetryRequiredPeers.has(PEER)).toBe(true);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('keeps the reserved selected-provider lane public-only and leaves private recovery ordinary', async () => {
-    const publicCg = 'selected-public-cg';
-    const privateCg = 'did:dkg:agent:0x1111111111111111111111111111111111111111/private-cg';
-    const syncCalls: Array<{
-      contextGraphIds: readonly string[];
-      selected: boolean;
-    }> = [];
-    const mixedPlan = {
-      publicContextGraphIds: [publicCg],
-      privateRecoverFromCurator: [privateCg],
-      eligibleContextGraphIds: [publicCg, privateCg],
-    };
-    const agent: SelectedProviderSelectionAgent = {
-      started: true,
-      config: {
-        syncOnConnect: true,
-        syncSharedMemoryOnConnect: true,
-        syncContextGraphs: [publicCg, privateCg],
-        rfc64PublicCatalogBootstrap: {
-          acceptedPublicPolicies: [{ completeSwmProviders: [PEER] }],
-        },
-      },
-      networkAdmissionCoordinator: { isAcceptedPeer: () => true },
-      syncingPeers: new Set<string>(),
-      knownCorePeerIds: new Set<string>(),
-      knownCorePeerIdsV2: new Set<string>(),
-      skippedNoSyncPeers: new Set<string>(),
-      lastSuccessfulSyncAt: new Map<string, number>(),
-      lastSyncProgressAt: new Map<string, number>(),
-      syncReconcilerBackoff: new Map<string, unknown>(),
-      selectedSwmRetryRequiredPeers: new Set<string>(),
-      getPeerProtocols: async () => [PROTOCOL_SYNC],
-      planSharedMemorySyncContextGraphs: async () => mixedPlan,
-      resolveRfc64CompleteSwmProviderPeerIdsV1: (contextGraphId: string) => (
-        contextGraphId === publicCg ? [PEER] : []
-      ),
-      syncFromPeerDetailed: async () => 0,
-      refreshMetaSyncedFlags: async () => undefined,
-      discoverContextGraphsFromStore: async () => 0,
-      syncSharedMemoryFromPeerDetailed: async (
-        _peerId: string,
-        contextGraphIds: readonly string[],
-      ) => {
-        syncCalls.push({
-          contextGraphIds: [...contextGraphIds],
-          selected: false,
-        });
-        return cleanDurableResult();
-      },
-      syncSelectedSharedMemoryFromPeerDetailed: async (
-        _peerId: string,
-        contextGraphIds: readonly string[],
-      ) => {
-        syncCalls.push({
-          contextGraphIds: [...contextGraphIds],
-          selected: true,
-        });
-        return {
-          kind: 'selected-shared-memory',
-          shared: cleanDurableResult(),
-          selectedScopeComplete: true,
-        };
-      },
-      log: { info: () => {}, warn: () => {}, debug: () => {} },
-    };
-
-    await callTrySyncFromPeer.call(agent, PEER);
-
-    expect(syncCalls).toEqual([
-      { contextGraphIds: [publicCg], selected: true },
-      { contextGraphIds: [privateCg], selected: false },
-    ]);
-  });
-
-  it('does not stamp a max-pass incomplete selected provider fresh', async () => {
-    const publicCg = 'selected-max-passes-incomplete';
-    const accounting: Array<{ fresh: boolean; progress?: boolean }> = [];
-    const agent: SelectedProviderSelectionAgent = {
-      started: true,
-      config: {
-        syncOnConnect: true,
-        syncSharedMemoryOnConnect: true,
-        syncContextGraphs: [publicCg],
-        rfc64PublicCatalogBootstrap: {
-          acceptedPublicPolicies: [{ completeSwmProviders: [PEER] }],
-        },
-      },
-      networkAdmissionCoordinator: { isAcceptedPeer: () => true },
-      syncingPeers: new Set<string>(),
-      knownCorePeerIds: new Set<string>(),
-      knownCorePeerIdsV2: new Set<string>(),
-      skippedNoSyncPeers: new Set<string>(),
-      lastSuccessfulSyncAt: new Map<string, number>(),
-      lastSyncProgressAt: new Map<string, number>(),
-      syncReconcilerBackoff: new Map<string, unknown>(),
-      selectedSwmRetryRequiredPeers: new Set<string>(),
-      getPeerProtocols: async () => [PROTOCOL_SYNC],
-      planSharedMemorySyncContextGraphs: async () => ({
-        publicContextGraphIds: [publicCg],
-        privateRecoverFromCurator: [],
-        eligibleContextGraphIds: [publicCg],
-      }),
-      resolveRfc64CompleteSwmProviderPeerIdsV1: () => [PEER],
-      syncFromPeerDetailed: async () => 0,
-      refreshMetaSyncedFlags: async () => undefined,
-      discoverContextGraphsFromStore: async () => 0,
-      syncSharedMemoryFromPeerDetailed: async () => cleanDurableResult(),
-      syncSelectedSharedMemoryFromPeerDetailed: async (
-        _peerId: string,
-        _contextGraphIds: readonly string[],
-      ) => ({
-        kind: 'selected-shared-memory',
-        shared: cleanDurableResult(),
-        // The producer carries whole-selected-scope completeness explicitly;
-        // freshness accounting must not depend on ambient peer state.
-        selectedScopeComplete: false,
-      }),
-      log: { info: () => {}, warn: () => {}, debug: () => {} },
-    };
-
-    await callTrySyncFromPeer.call(agent, PEER, (outcome) => accounting.push(outcome));
-
-    expect(agent.selectedSwmRetryRequiredPeers.has(PEER)).toBe(false);
-    expect(agent.lastSuccessfulSyncAt.has(PEER)).toBe(false);
-    // No freshness/progress callback means the reconciler wrapper classifies
-    // this explicit incomplete result as a failed attempt and grows backoff.
-    expect(accounting).toEqual([]);
-  });
-
-  it('continues a voluntary 672/905 public snapshot yield to 905/905 with fresh admission', async () => {
-    const publicCg = 'selected-production-shape';
-    const privateCg = 'selected-private-unaffected';
-    let initialSnapshotReads = 0;
-    let wallNow = 1_000;
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg, private: privateCg },
-      manifest: snapshotManifest(publicCg, 905),
-      clock: { now: () => wallNow, deadline: () => wallNow + 1 },
-      priorities: { [publicCg]: 100, [privateCg]: 0 },
-      onSnapshotRead: ({ publicAdmission }) => {
-        if (publicAdmission === 1) {
-          initialSnapshotReads += 1;
-          if (initialSnapshotReads === 672) wallNow += 120_000;
-        }
-      },
-    });
-
-    try {
-      const selected = await callSelectedSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg, privateCg],
-        {
-        selectedSwmPriority: true,
-        priority: 2_000,
-        sharedMemorySyncPlan: {
-          publicContextGraphIds: [publicCg],
-          privateRecoverFromCurator: [privateCg],
-          eligibleContextGraphIds: [publicCg, privateCg],
-        },
-        },
-      );
-      const summary = selected.shared;
-
-      expect(summary.swmCoverage).toMatchObject({
-        contextGraphId: publicCg,
-        snapshotsResolved: 905,
-        snapshotsTotal: 905,
-        missingCount: 0,
-      });
-      expect(selected.selectedScopeComplete).toBe(true);
-      expect(summary.continuationPasses).toBe(1);
-      expect(summary.snapshotPlaneIncomplete).toBe(1);
-      expect(summary.failedPhases).toBe(1);
-      expect(summary.resolvedSnapshotPlaneIncomplete).toBe(1);
-      expect(summary.timedOutPhases).toBe(0);
-      expect(summary.backoffWorthyFailures).toBe(0);
-      expect(harness.agent.selectedSwmRetryRequiredPeers.has(PEER)).toBe(false);
-      expect(initialSnapshotReads).toBe(672);
-      expect(harness.probes.snapshotFetches).toEqual([]);
-      expect(harness.probes.maxActiveAdmissions()).toBe(1);
-      expect(harness.probes.admissions).toEqual([
-        { contextGraphId: publicCg, selected: true, priority: 2_000, event: 'start' },
-        { contextGraphId: publicCg, selected: true, priority: 2_000, event: 'end' },
-        { contextGraphId: privateCg, selected: false, priority: undefined, event: 'start' },
-        { contextGraphId: privateCg, selected: false, priority: undefined, event: 'end' },
-        { contextGraphId: publicCg, selected: true, priority: 2_000, event: 'start' },
-        { contextGraphId: publicCg, selected: true, priority: 2_000, event: 'end' },
-      ]);
-
-      const onPeerSynced = vi.fn();
-      const outcome = await runSyncOnConnect({
-        remotePeer: PEER,
-        syncingPeers: new Set(),
-        getPeerProtocols: async () => [PROTOCOL_SYNC],
-        knownCorePeerIds: new Set(),
-        knownCorePeerIdsV2: new Set(),
-        getSyncContextGraphs: () => [],
-        getDurableSyncContextGraphs: () => [],
-        getSharedMemorySyncContextGraphs: () => [publicCg],
-        selectedSharedMemoryLane: {
-          getContextGraphIds: () => [publicCg],
-          syncFromPeer: async () => selected,
-        },
-        syncFromPeer: async () => 0,
-        refreshMetaSyncedFlags: async () => undefined,
-        discoverContextGraphsFromStore: async () => 0,
-        syncSharedMemoryFromPeer: async () => summary,
-        onPeerSynced,
-        logInfo: () => {},
-      });
-      expect(outcome).toBe('synced');
-      expect(onPeerSynced).toHaveBeenCalledWith(PEER, {
-        fresh: true,
-        progress: false,
-      });
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('keeps retry state when one selected CG completes and another remains incomplete', async () => {
-    const completeCg = 'selected-multi-complete';
-    const incompleteCg = 'selected-multi-incomplete';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: completeCg },
-      manifest: snapshotManifest(completeCg, 2),
-      clock: { now: () => 1_000, deadline: () => 61_000 },
-    });
-
-    try {
-      const selected = await callSelectedSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [completeCg, incompleteCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [completeCg, incompleteCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [completeCg, incompleteCg],
-          },
-        },
-      );
-      const summary = selected.shared;
-
-      expect(selected.selectedScopeComplete).toBe(false);
-      expect(harness.agent.selectedSwmRetryRequiredPeers.has(PEER)).toBe(true);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('retains an exact metadata prefix across bounded passes and verifies the combined manifest once', async () => {
-    const publicCg = 'selected-multi-pass-metadata';
-    const manifest = snapshotManifest(publicCg, 3);
-    const firstPrefix = manifest.meta.slice(0, 4);
-    const finalSuffix = manifest.meta.slice(4);
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaPages: [
-        {
-          quads: firstPrefix,
-          resumedFromOffset: 0,
-          nextOffset: firstPrefix.length,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: finalSuffix,
-          resumedFromOffset: firstPrefix.length,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-    });
-
-    try {
-      const summary = await callSyncSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          stopOnBackoffWorthyFailure: true,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(harness.probes.metaFetches()).toBe(2);
-      expect(harness.probes.metaRequesterScopes).toHaveLength(2);
-      expect(harness.probes.metaRequesterScopes[0]).toMatch(/^selected-swm-meta:retained:\d+$/);
-      expect(harness.probes.metaRequesterScopes[1]).toBe(
-        harness.probes.metaRequesterScopes[0],
-      );
-      expect(harness.probes.metaSinceBatchIds).toEqual([undefined, undefined]);
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-      expect(harness.probes.publicAdmissions()).toBe(2);
-      expect(summary.continuationPasses).toBe(1);
-      expect(summary.metadataContinuationYields).toBe(1);
-      expect(summary.timedOutPhases).toBe(1);
-      expect(summary.resolvedMetadataContinuationYields).toBe(1);
-      expect(classifySharedMemoryFreshness(summary).timedOut).toBe(false);
-      expect(summary.swmCoverage).toMatchObject({
-        contextGraphId: publicCg,
-        snapshotsResolved: 3,
-        snapshotsTotal: 3,
-        missingCount: 0,
-      });
-      expect(harness.agent.syncCheckpoints.size).toBe(0);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('resumes one exact metadata prefix across outer reconciler invocations without partial activation', async () => {
-    const publicCg = 'selected-cross-outer-metadata';
-    const manifest = snapshotManifest(publicCg, 3);
-    const firstPrefix = manifest.meta.slice(0, 4);
-    const finalSuffix = manifest.meta.slice(4);
-    const previousBudget = process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-    process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = '0';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaPages: [
-        {
-          quads: firstPrefix,
-          resumedFromOffset: 0,
-          nextOffset: firstPrefix.length,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: finalSuffix,
-          resumedFromOffset: firstPrefix.length,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-    });
-    const plan = {
-      publicContextGraphIds: [publicCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [publicCg],
-    };
-
-    try {
-      const firstSelected = await callSelectedSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        { selectedSwmPriority: true, priority: 2_000, sharedMemorySyncPlan: plan },
-      );
-      const first = firstSelected.shared;
-
-      expect(first.metadataContinuationYields).toBe(1);
-      expect(firstSelected.selectedScopeComplete).toBe(false);
-      expect(harness.probes.processedMetaBatches).toEqual([]);
-      expect(harness.agent.selectedSwmRetryRequiredPeers.has(PEER)).toBe(true);
-
-      const queuedPeers: string[] = [];
-      const queueAgent = harness.agent as SelectedSwmLifecycleAgentFixture & Record<string, any>;
-      queueAgent.networkAdmissionCoordinator = { isAcceptedPeer: () => true };
-      queueAgent.lastSuccessfulSyncAt = new Map([[PEER, Date.now()]]);
-      queueAgent.lastSyncDisconnectedAt = new Map<string, number>();
-      queueAgent.catchupOnConnectAt = new Map<string, number>();
-      queueAgent.syncReconcilerBackoff = new Map<string, unknown>();
-      queueAgent.syncOnConnectDisconnectBoundary =
-        LifecycleSyncMethods.prototype.syncOnConnectDisconnectBoundary;
-      queueAgent.runSyncFromPeerOnConnect = async (peerId: string) => {
-        queuedPeers.push(peerId);
-      };
-      expect(LifecycleSyncMethods.prototype.queueSyncFromPeerOnConnect.call(
-        queueAgent as never,
-        PEER,
-        () => undefined,
-        0,
-        { selectedSwmRetry: true },
-      )).toBe(true);
-      await new Promise((resolve) => setTimeout(resolve, 0));
-      expect(queuedPeers).toEqual([PEER]);
-
-      const secondSelected = await callSelectedSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        { selectedSwmPriority: true, priority: 2_000, sharedMemorySyncPlan: plan },
-      );
-      const second = secondSelected.shared;
-
-      expect(second.failedPhases).toBe(0);
-      expect(secondSelected.selectedScopeComplete).toBe(true);
-      expect(harness.probes.metaRequesterScopes).toHaveLength(2);
-      expect(harness.probes.metaRequesterScopes[1]).toBe(
-        harness.probes.metaRequesterScopes[0],
-      );
-      expect(
-        harness.probes.metaReturnAcceptedPrefixOnRetryableTransportFailure,
-      ).toEqual([true, true]);
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-      expect(harness.agent.selectedSwmRetryRequiredPeers.has(PEER)).toBe(false);
-    } finally {
-      if (previousBudget === undefined) {
-        delete process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-      } else {
-        process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = previousBudget;
-      }
-      await harness.close();
-    }
-  });
-
-  it('expires a cross-invocation prefix and starts a fresh responder scope', async () => {
-    const publicCg = 'selected-cross-outer-expiry';
-    const manifest = snapshotManifest(publicCg, 2);
-    let wallNow = 1_000;
-    const previousBudget = process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-    process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = '0';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => wallNow, deadline: () => wallNow + 1 },
-      metaPages: [
-        {
-          quads: manifest.meta.slice(0, 2),
-          resumedFromOffset: 0,
-          nextOffset: 2,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-    });
-    const plan = {
-      publicContextGraphIds: [publicCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [publicCg],
-    };
-
-    try {
-      await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        sharedMemorySyncPlan: plan,
-      });
-      wallNow += DURABLE_DATA_SYNC_SESSION_TTL_MS + 1;
-      await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        sharedMemorySyncPlan: plan,
-      });
-
-      expect(harness.probes.metaRequesterScopes).toHaveLength(2);
-      expect(harness.probes.metaRequesterScopes[1]).not.toBe(
-        harness.probes.metaRequesterScopes[0],
-      );
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-    } finally {
-      if (previousBudget === undefined) {
-        delete process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-      } else {
-        process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = previousBudget;
-      }
-      await harness.close();
-    }
-  });
-
-  it('drops an old cross-invocation prefix when the responder replaces its session', async () => {
-    const publicCg = 'selected-cross-outer-session-replacement';
-    const manifest = snapshotManifest(publicCg, 2);
-    const previousBudget = process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-    process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = '0';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaPages: [
-        {
-          quads: manifest.meta.slice(0, 2),
-          resumedFromOffset: 0,
-          nextOffset: 14_000,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-    });
-    const plan = {
-      publicContextGraphIds: [publicCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [publicCg],
-    };
-
-    try {
-      await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        sharedMemorySyncPlan: plan,
-      });
-      await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        sharedMemorySyncPlan: plan,
-      });
-
-      expect(harness.probes.metaRequesterScopes[1]).toBe(
-        harness.probes.metaRequesterScopes[0],
-      );
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-    } finally {
-      if (previousBudget === undefined) {
-        delete process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-      } else {
-        process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = previousBudget;
-      }
-      await harness.close();
-    }
-  });
-
-  it.each([
-    ['caller abort', Object.assign(new Error('caller cancelled'), { name: 'AbortError' })],
-    ['integrity rejection', new Error('metadata integrity rejected')],
-    ['local request build failure', new Error('wallet signing failed')],
-  ])('discards retained state after a cross-invocation %s', async (_label, failure) => {
-    const publicCg = `selected-cross-outer-failure-${_label.replaceAll(' ', '-')}`;
-    const manifest = snapshotManifest(publicCg, 2);
-    const previousBudget = process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-    process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = '0';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaPages: [
-        {
-          quads: manifest.meta.slice(0, 2),
-          resumedFromOffset: 0,
-          nextOffset: 2,
-          completed: false,
-          timedOut: true,
-        },
-        // Fetch two throws from onMetaFetch. Fetch three must therefore begin
-        // a new owner at offset zero and may accept this full replacement.
-        {
-          quads: [],
-          resumedFromOffset: 0,
-          nextOffset: 0,
-          completed: false,
-          timedOut: false,
-        },
-        {
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-      onMetaFetch: ({ fetch }) => {
-        if (fetch === 2) throw failure;
-      },
-    });
-    const plan = {
-      publicContextGraphIds: [publicCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [publicCg],
-    };
-
-    try {
-      await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        sharedMemorySyncPlan: plan,
-      });
-      const failed = await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        sharedMemorySyncPlan: plan,
-      });
-      expect(failed.failedPhases + failed.failedPeers).toBeGreaterThan(0);
-      await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        sharedMemorySyncPlan: plan,
-      });
-
-      expect(harness.probes.metaRequesterScopes).toHaveLength(3);
-      expect(harness.probes.metaRequesterScopes[1]).toBe(
-        harness.probes.metaRequesterScopes[0],
-      );
-      expect(harness.probes.metaRequesterScopes[2]).not.toBe(
-        harness.probes.metaRequesterScopes[0],
-      );
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-    } finally {
-      if (previousBudget === undefined) {
-        delete process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-      } else {
-        process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = previousBudget;
-      }
-      await harness.close();
-    }
-  });
-
-  it('releases retained prefixes on node-lifecycle cleanup', async () => {
-    const publicCg = 'selected-cross-outer-node-close';
-    const manifest = snapshotManifest(publicCg, 2);
-    const previousBudget = process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-    process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = '0';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaPages: [
-        {
-          quads: manifest.meta.slice(0, 2),
-          resumedFromOffset: 0,
-          nextOffset: 2,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-    });
-    const plan = {
-      publicContextGraphIds: [publicCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [publicCg],
-    };
-
-    try {
-      await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        sharedMemorySyncPlan: plan,
-      });
-      await harness.agent.closeSelectedSwmMetaTransfers();
-      await callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        sharedMemorySyncPlan: plan,
-      });
-
-      expect(harness.probes.metaRequesterScopes).toHaveLength(2);
-      expect(harness.probes.metaRequesterScopes[1]).not.toBe(
-        harness.probes.metaRequesterScopes[0],
-      );
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-    } finally {
-      if (previousBudget === undefined) {
-        delete process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-      } else {
-        process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = previousBudget;
-      }
-      await harness.close();
-    }
-  });
-
-  it('resolves a completed metadata continuation while a later snapshot yield remains active', async () => {
-    const publicCg = 'selected-metadata-complete-snapshot-incomplete';
-    const manifest = snapshotManifest(publicCg, 2);
-    const firstPrefix = manifest.meta.slice(0, 3);
-    const finalSuffix = manifest.meta.slice(3);
-    let wallNow = 1_000;
-    const previousMaxPasses = process.env.DKG_SWM_CATCHUP_MAX_PASSES;
-    process.env.DKG_SWM_CATCHUP_MAX_PASSES = '2';
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => wallNow, deadline: () => wallNow + 1 },
-      metaPages: [
-        {
-          quads: firstPrefix,
-          resumedFromOffset: 0,
-          nextOffset: firstPrefix.length,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: finalSuffix,
-          resumedFromOffset: firstPrefix.length,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-      onSnapshotRead: ({ publicAdmission, snapshotRead }) => {
-        if (publicAdmission === 2 && snapshotRead === 1) wallNow += 120_000;
-      },
-    });
-
-    try {
-      const summary = await callSyncSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(summary.continuationPasses).toBe(1);
-      expect(summary.metadataContinuationYields).toBe(1);
-      expect(summary.timedOutPhases).toBe(1);
-      expect(summary.resolvedMetadataContinuationYields).toBe(1);
-      expect(summary.snapshotPlaneIncomplete).toBe(1);
-      expect(summary.failedPhases).toBe(1);
-      expect(summary.resolvedSnapshotPlaneIncomplete).toBe(0);
-      const freshness = classifySharedMemoryFreshness(summary);
-      expect(freshness.timedOut).toBe(false);
-      expect(freshness.backoffWorthyFailure).toBe(false);
-      expect(freshness.phaseFailed).toBe(true);
-    } finally {
-      if (previousMaxPasses === undefined) {
-        delete process.env.DKG_SWM_CATCHUP_MAX_PASSES;
-      } else {
-        process.env.DKG_SWM_CATCHUP_MAX_PASSES = previousMaxPasses;
-      }
-      await harness.close();
-    }
-  });
-
-  it('grants the newly capable snapshot domain a pass after an empty metadata EOF page', async () => {
-    const publicCg = 'selected-empty-meta-eof-before-snapshots';
-    const manifest = snapshotManifest(publicCg, 3);
-    let wallNow = 1_000;
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => wallNow, deadline: () => wallNow + 1 },
-      metaPages: [
-        {
-          // The first pass retained the entire manifest, but did not receive
-          // the responder EOF marker before its bounded deadline.
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: manifest.meta.length,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          // Exact empty EOF tail. Metadata is now complete, but this pass has
-          // no time left to begin the newly discovered snapshot domain.
-          quads: [],
-          resumedFromOffset: manifest.meta.length,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-      onMetaFetch: ({ fetch }) => {
-        if (fetch === 2) wallNow += 120_000;
-      },
-    });
-
-    try {
-      const summary = await callSyncSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(harness.probes.metaFetches()).toBe(2);
-      expect(harness.probes.publicAdmissions()).toBe(3);
-      expect(harness.probes.snapshotReads()).toBe(3);
-      expect(summary.continuationPasses).toBe(2);
-      expect(summary.metadataContinuationYields).toBe(1);
-      expect(summary.resolvedMetadataContinuationYields).toBe(1);
-      expect(summary.snapshotPlaneIncomplete).toBe(1);
-      expect(summary.resolvedSnapshotPlaneIncomplete).toBe(1);
-      expect(summary.swmCoverage).toMatchObject({
-        snapshotsResolved: 3,
-        snapshotsTotal: 3,
-        missingCount: 0,
-      });
-      expect(classifySharedMemoryFreshness(summary)).toMatchObject({
-        phaseFailed: false,
-        timedOut: false,
-        backoffWorthyFailure: false,
-      });
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('treats a responder restart as a new progress generation and completes its replacement prefix', async () => {
-    const publicCg = 'selected-restarted-metadata-generation';
-    const manifest = snapshotManifest(publicCg, 3);
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaPages: [
-        {
-          quads: manifest.meta.slice(0, 2),
-          resumedFromOffset: 0,
-          nextOffset: 14_000,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: manifest.meta.slice(0, 2),
-          resumedFromOffset: 0,
-          nextOffset: 5_000,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: manifest.meta.slice(2),
-          resumedFromOffset: 5_000,
-          nextOffset: 19_000,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-    });
-
-    try {
-      const summary = await callSyncSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(harness.probes.metaFetches()).toBe(3);
-      expect(harness.probes.publicAdmissions()).toBe(3);
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-      expect(summary.continuationPasses).toBe(2);
-      expect(summary.metadataContinuationYields).toBe(2);
-      expect(summary.resolvedMetadataContinuationYields).toBe(2);
-      expect(summary.swmCoverage).toMatchObject({
-        snapshotsResolved: 3,
-        snapshotsTotal: 3,
-        missingCount: 0,
-      });
-      expect(harness.agent.syncCheckpoints.size).toBe(0);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('drops a full old prefix and retries once when a fresh responder generation replaces it', async () => {
-    const publicCg = 'selected-full-prefix-fresh-generation';
-    const manifest = snapshotManifest(publicCg, 1);
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaContinuationLimits: { rows: manifest.meta.length, bytesEstimate: 1024 * 1024 },
-      metaPages: [
-        {
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: 14_000,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          // The old prefix owns the entire row allowance, so this first fresh
-          // replacement attempt proves the restart by hitting the zero append
-          // cap. The requester must discard the old generation and retry.
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: 2,
-          completed: true,
-          timedOut: false,
-          responderSessionStartedFresh: true,
-        },
-        {
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: 2,
-          completed: true,
-          timedOut: false,
-          responderSessionStartedFresh: true,
-        },
-      ],
-    });
-
-    try {
-      const summary = await callSyncSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(harness.probes.metaFetches()).toBe(3);
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-      expect(summary.continuationPasses).toBe(1);
-      expect(summary.metadataContinuationYields).toBe(1);
-      expect(summary.resolvedMetadataContinuationYields).toBe(1);
-      expect(summary.failedPhases).toBe(0);
-      expect(harness.agent.syncCheckpoints.size).toBe(0);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('serializes overlapping selected invocations behind one retained-prefix owner', async () => {
-    const publicCg = 'selected-overlap-isolation';
-    const manifest = snapshotManifest(publicCg, 2);
-    const firstPrefix = manifest.meta.slice(0, 2);
-    const finalSuffix = manifest.meta.slice(2);
-    const previousBudget = process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-    process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = '0';
-    let signalFirstStarted!: () => void;
-    const firstStarted = new Promise<void>((resolve) => { signalFirstStarted = resolve; });
-    let releaseFirst!: () => void;
-    const firstRelease = new Promise<void>((resolve) => { releaseFirst = resolve; });
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 61_000 },
-      metaPages: [
-        {
-          quads: firstPrefix,
-          resumedFromOffset: 0,
-          nextOffset: firstPrefix.length,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: finalSuffix,
-          resumedFromOffset: firstPrefix.length,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-      onMetaFetch: async ({ fetch }) => {
-        if (fetch === 1) {
-          signalFirstStarted();
-          await firstRelease;
-        }
-      },
-    });
-    const plan = {
-      publicContextGraphIds: [publicCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [publicCg],
-    };
-
-    try {
-      const first = callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        priority: 2_000,
-        sharedMemorySyncPlan: plan,
-      });
-      await firstStarted;
-      const second = callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        priority: 2_001,
-        sharedMemorySyncPlan: plan,
-      });
-      await Promise.resolve();
-      expect(harness.probes.metaFetches()).toBe(1);
-      releaseFirst();
-      await Promise.all([first, second]);
-
-      expect(harness.probes.metaRequesterScopes).toHaveLength(2);
-      expect(new Set(harness.probes.metaRequesterScopes).size).toBe(1);
-      expect(harness.probes.metaSinceBatchIds).toEqual([undefined, undefined]);
-      expect(harness.probes.maxActiveAdmissions()).toBe(1);
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-      expect(harness.agent.syncCheckpoints.size).toBe(0);
-    } finally {
-      releaseFirst();
-      if (previousBudget === undefined) {
-        delete process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-      } else {
-        process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = previousBudget;
-      }
-      await harness.close();
-    }
-  });
-
-  it('does not coalesce selected SWM into an ordinary same-priority flight', async () => {
-    const publicCg = 'selected-ordinary-single-flight-isolation';
-    const manifest = snapshotManifest(publicCg, 2);
-    const firstPrefix = manifest.meta.slice(0, 3);
-    const finalSuffix = manifest.meta.slice(3);
-    let signalOrdinaryStarted!: () => void;
-    const ordinaryStarted = new Promise<void>((resolve) => {
-      signalOrdinaryStarted = resolve;
-    });
-    let releaseOrdinary!: () => void;
-    const ordinaryRelease = new Promise<void>((resolve) => {
-      releaseOrdinary = resolve;
-    });
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaPages: [
-        {
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-        {
-          quads: firstPrefix,
-          resumedFromOffset: 0,
-          nextOffset: firstPrefix.length,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: finalSuffix,
-          resumedFromOffset: firstPrefix.length,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-      onMetaFetch: async ({ fetch }) => {
-        if (fetch === 1) {
-          signalOrdinaryStarted();
-          await ordinaryRelease;
-        }
-      },
-    });
-    const plan = {
-      publicContextGraphIds: [publicCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [publicCg],
-    };
-    let ordinary: Promise<SharedMemorySyncResult> | undefined;
-    let selected: Promise<SharedMemorySyncResult> | undefined;
-
-    try {
-      ordinary = callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        priority: 2_000,
-        sharedMemorySyncPlan: plan,
-      });
-      await ordinaryStarted;
-
-      selected = callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        priority: 2_000,
-        sharedMemorySyncPlan: plan,
-      });
-      await vi.waitFor(() => expect(harness.probes.metaFetches()).toBeGreaterThanOrEqual(2));
-      releaseOrdinary();
-
-      const [ordinarySummary, selectedSummary] = await Promise.all([ordinary, selected]);
-      expect(ordinarySummary).not.toBe(selectedSummary);
-      expect(harness.probes.metaFetches()).toBe(3);
-      expect(harness.probes.metaRequesterScopes[0]).toBeUndefined();
-      expect(harness.probes.metaRequesterScopes[1]).toMatch(/^selected-swm-meta:retained:\d+$/);
-      expect(harness.probes.metaRequesterScopes[2]).toBe(
-        harness.probes.metaRequesterScopes[1],
-      );
-      expect(selectedSummary.continuationPasses).toBe(1);
-      expect(selectedSummary.resolvedMetadataContinuationYields).toBe(1);
-      expect(selectedSummary.swmCoverage).toMatchObject({
-        contextGraphId: publicCg,
-        snapshotsResolved: 2,
-        snapshotsTotal: 2,
-        missingCount: 0,
-      });
-      expect(harness.agent.syncCheckpoints.size).toBe(0);
-    } finally {
-      releaseOrdinary();
-      await Promise.allSettled([ordinary, selected].filter(
-        (promise): promise is Promise<SharedMemorySyncResult> => promise !== undefined,
-      ));
-      await harness.close();
-    }
-  });
-
-  it('releases the shared metadata budget before the next serialized owner generation', async () => {
-    const publicCg = 'selected-overlap-shared-retention-budget';
-    const manifest = snapshotManifest(publicCg, 2);
-    let firstFetchStarted!: () => void;
-    const firstStarted = new Promise<void>((resolve) => { firstFetchStarted = resolve; });
-    let releaseFirst!: () => void;
-    const firstRelease = new Promise<void>((resolve) => { releaseFirst = resolve; });
-    const acceptedRows: Array<number | undefined> = [];
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 61_000 },
-      metaContinuationLimits: {
-        rows: manifest.meta.length,
-        bytesEstimate: 1024 * 1024,
-        globalRows: manifest.meta.length,
-        globalBytesEstimate: 1024 * 1024,
-      },
-      metaPages: [
-        {
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-        {
-          quads: manifest.meta,
-          resumedFromOffset: 0,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-      onMetaFetch: async ({ fetch, maxAcceptedQuads }) => {
-        acceptedRows.push(maxAcceptedQuads);
-        if (fetch === 1) {
-          firstFetchStarted();
-          await firstRelease;
-        }
-      },
-    });
-    const plan = {
-      publicContextGraphIds: [publicCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [publicCg],
-    };
-
-    try {
-      const first = callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        priority: 2_000,
-        sharedMemorySyncPlan: plan,
-      });
-      await firstStarted;
-      const second = callSyncSharedMemoryFromPeerDetailed(harness.agent, [publicCg], {
-        selectedSwmPriority: true,
-        priority: 2_001,
-        sharedMemorySyncPlan: plan,
-      });
-      await Promise.resolve();
-      expect(harness.probes.metaFetches()).toBe(1);
-      releaseFirst();
-
-      const [firstSummary, secondSummary] = await Promise.all([first, second]);
-
-      expect(acceptedRows).toEqual([manifest.meta.length, manifest.meta.length]);
-      expect(harness.probes.metaFetches()).toBe(2);
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta, manifest.meta]);
-      expect(firstSummary.failedPhases).toBe(0);
-      expect(secondSummary.failedPhases).toBe(0);
-      expect(harness.agent.syncCheckpoints.size).toBe(0);
-    } finally {
-      releaseFirst();
-      await harness.close();
-    }
-  });
-
-  it('does not lose a queued incomplete retry marker when an earlier call completes', async () => {
-    const completeCg = 'selected-overlap-complete';
-    const incompleteCg = 'selected-overlap-incomplete';
-    let wallNow = 1_000;
-    let signalFirstStarted!: () => void;
-    const firstStarted = new Promise<void>((resolve) => { signalFirstStarted = resolve; });
-    let releaseFirst!: () => void;
-    const firstRelease = new Promise<void>((resolve) => { releaseFirst = resolve; });
-    let firstBlocked = false;
-    const harness = createSelectedSwmLifecycleHarness({
-      // The first selected call is a clean empty 0/0 graph. The queued second
-      // call owns this two-ref manifest and yields after materializing one ref,
-      // so it remains genuinely incomplete after zero-ref completion became
-      // explicit terminal evidence.
-      contextGraphs: { public: incompleteCg },
-      manifest: snapshotManifest(incompleteCg, 2),
-      clock: { now: () => wallNow, deadline: () => wallNow + 1 },
-      reportEmptyResponse: true,
-      beforeAdmissionRun: async ({ contextGraphId }) => {
-        if (contextGraphId !== completeCg || firstBlocked) return;
-        firstBlocked = true;
-        signalFirstStarted();
-        await firstRelease;
-      },
-      onSnapshotRead: () => { wallNow += 120_000; },
-    });
-    const completePlan = {
-      publicContextGraphIds: [completeCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [completeCg],
-    };
-    const incompletePlan = {
-      publicContextGraphIds: [incompleteCg],
-      privateRecoverFromCurator: [],
-      eligibleContextGraphIds: [incompleteCg],
-    };
-
-    try {
-      const first = callSyncSharedMemoryFromPeerDetailed(harness.agent, [completeCg], {
-        selectedSwmPriority: true,
-        priority: 2_000,
-        sharedMemorySyncPlan: completePlan,
-      });
-      await firstStarted;
-      const second = callSyncSharedMemoryFromPeerDetailed(harness.agent, [incompleteCg], {
-        selectedSwmPriority: true,
-        priority: 2_001,
-        sharedMemorySyncPlan: incompletePlan,
-      });
-      await Promise.resolve();
-      releaseFirst();
-
-      const [completeSummary, incompleteSummary] = await Promise.all([first, second]);
-      expect(completeSummary.swmCoverage).toMatchObject({
-        contextGraphId: completeCg,
-        snapshotsResolved: 0,
-        snapshotsTotal: 0,
-      });
-      expect(incompleteSummary.swmCoverage).toMatchObject({
-        contextGraphId: incompleteCg,
-        snapshotsResolved: 1,
-        snapshotsTotal: 2,
-      });
-      expect(harness.agent.selectedSwmRetryRequiredPeers.has(PEER)).toBe(true);
-    } finally {
-      releaseFirst();
-      await harness.close();
-    }
-  });
-
-  it('fails closed when a retained metadata prefix exceeds the local resource budget', async () => {
-    const publicCg = 'selected-metadata-resource-cap';
-    const manifest = snapshotManifest(publicCg, 2);
-    const onMetaFetch = vi.fn();
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaContinuationLimits: { rows: 1, bytesEstimate: 1024 * 1024 },
-      onMetaFetch,
-      metaPages: [{
-        quads: manifest.meta.slice(0, 2),
-        resumedFromOffset: 0,
-        nextOffset: 2,
-        completed: false,
-        timedOut: true,
-      }],
-    });
-
-    try {
-      const summary = await callSyncSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(harness.probes.metaFetches()).toBe(1);
-      expect(onMetaFetch).toHaveBeenCalledWith(expect.objectContaining({
-        maxAcceptedQuads: 1,
-        maxAcceptedHeapBytesEstimate: 1024 * 1024,
-      }));
-      expect(harness.probes.processedMetaBatches).toEqual([]);
-      expect(summary.failedPhases).toBe(1);
-      expect(summary.continuationPasses).toBe(0);
-      expect(harness.agent.syncCheckpoints.size).toBe(0);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('reuses a complete retained manifest when a later aggregate-data phase fails', async () => {
-    const publicCg = 'selected-complete-metadata-reuse';
-    const manifest = snapshotManifest(publicCg, 3);
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 61_000 },
-      dataFailuresBeforeSuccess: 1,
-    });
-
-    try {
-      const summary = await callSyncSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(harness.probes.publicAdmissions()).toBe(2);
-      expect(harness.probes.metaFetches()).toBe(1);
-      expect(harness.probes.dataFetches()).toBe(2);
-      expect(harness.probes.processedMetaBatches).toEqual([manifest.meta]);
-      expect(summary.continuationPasses).toBe(1);
-      expect(summary.swmCoverage).toMatchObject({
-        snapshotsResolved: 3,
-        snapshotsTotal: 3,
-      });
-      expect(harness.agent.syncCheckpoints.size).toBe(0);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('fails closed instead of splicing a tail from an incompatible metadata offset', async () => {
-    const publicCg = 'selected-metadata-offset-mismatch';
-    const manifest = snapshotManifest(publicCg, 3);
-    const firstPrefix = manifest.meta.slice(0, 4);
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: publicCg },
-      manifest,
-      clock: { now: () => 1_000, deadline: () => 1_001 },
-      metaPages: [
-        {
-          quads: firstPrefix,
-          resumedFromOffset: 0,
-          nextOffset: firstPrefix.length,
-          completed: false,
-          timedOut: true,
-        },
-        {
-          quads: manifest.meta.slice(1),
-          resumedFromOffset: 1,
-          nextOffset: manifest.meta.length,
-          completed: true,
-          timedOut: false,
-        },
-      ],
-    });
-
-    try {
-      const summary = await callSyncSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [publicCg],
-        {
-          selectedSwmPriority: true,
-          priority: 2_000,
-          stopOnBackoffWorthyFailure: true,
-          sharedMemorySyncPlan: {
-            publicContextGraphIds: [publicCg],
-            privateRecoverFromCurator: [],
-            eligibleContextGraphIds: [publicCg],
-          },
-        },
-      );
-
-      expect(harness.probes.metaFetches()).toBe(2);
-      expect(harness.probes.processedMetaBatches).toEqual([]);
-      expect(summary.failedPhases).toBe(1);
-      expect(summary.swmCoverage).toBeUndefined();
-      expect(harness.agent.syncCheckpoints.size).toBe(0);
-    } finally {
-      await harness.close();
-    }
-  });
-
-  it('rechecks the continuation budget after queued scheduler admission', async () => {
-    const contextGraphId = 'selected-expired-in-queue';
-    const priorBudget = process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-    process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = '50';
-    let wallNow = 2_000;
-    const harness = createSelectedSwmLifecycleHarness({
-      contextGraphs: { public: contextGraphId },
-      manifest: snapshotManifest(contextGraphId, 2),
-      clock: { now: () => wallNow, deadline: () => wallNow + 1 },
-      onSnapshotRead: ({ publicAdmission, snapshotRead }) => {
-        if (publicAdmission === 1 && snapshotRead === 1) wallNow += 120_000;
-      },
-      beforeAdmissionRun: async ({ publicAdmission }) => {
-        if (publicAdmission === 2) {
-          await new Promise((resolve) => setTimeout(resolve, 75));
-        }
-      },
-    });
-
-    try {
-      const summary = await callSyncSharedMemoryFromPeerDetailed(
-        harness.agent,
-        [contextGraphId],
-        {
-        selectedSwmPriority: true,
-        priority: 2_000,
-        sharedMemorySyncPlan: {
-          publicContextGraphIds: [contextGraphId],
-          privateRecoverFromCurator: [],
-          eligibleContextGraphIds: [contextGraphId],
-        },
-        },
-      );
-
-      expect(harness.probes.publicAdmissions()).toBe(2);
-      expect(harness.probes.metaFetches()).toBe(1);
-      expect(harness.probes.snapshotReads()).toBe(1);
-      expect(summary.continuationPasses).toBe(0);
-      expect(summary.swmCoverage).toMatchObject({
-        snapshotsResolved: 1,
-        snapshotsTotal: 2,
-      });
-    } finally {
-      if (priorBudget === undefined) delete process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS;
-      else process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS = priorBudget;
-      await harness.close();
-    }
-  });
-
-  it('does not hide an unrelated shared-memory phase failure', async () => {
-    const contextGraphId = 'selected-with-unrelated-failure';
-    const onPeerSynced = vi.fn();
-    const shared = {
-      ...result(contextGraphId, 3, 3),
-      failedPhases: 2,
-      snapshotPlaneIncomplete: 1,
-      resolvedSnapshotPlaneIncomplete: 1,
-    };
-
-    await runSyncOnConnect({
-      remotePeer: PEER,
-      syncingPeers: new Set(),
-      getPeerProtocols: async () => [PROTOCOL_SYNC],
-      knownCorePeerIds: new Set(),
-      knownCorePeerIdsV2: new Set(),
-      getSyncContextGraphs: () => [],
-      getDurableSyncContextGraphs: () => [],
-      getSharedMemorySyncContextGraphs: () => [contextGraphId],
-      selectedSharedMemoryLane: {
-        getContextGraphIds: () => [contextGraphId],
-        syncFromPeer: async () => ({
-          kind: 'selected-shared-memory',
-          shared,
-          selectedScopeComplete: true,
-        }),
-      },
-      syncFromPeer: async () => 0,
-      refreshMetaSyncedFlags: async () => undefined,
-      discoverContextGraphsFromStore: async () => 0,
-      syncSharedMemoryFromPeer: async () => shared,
-      onPeerSynced,
-      logInfo: () => {},
-    });
-
-    expect(onPeerSynced).toHaveBeenCalledWith(PEER, {
-      fresh: false,
-      progress: true,
-    });
   });
 });
